Door opening/closing & locking logic
Doors now open/close in a way that makes sense when overlap is triggered. (Will make it check if player specially set it off soon but may include enemies too not 100% yet).
Doors are managed by a "DoorManager" object which has a list of doors in the scene (manually set) and can access any door in list to unlock.
Changed Character from abstract class to interface due to inheritance only enabling 1 class and an interface fits the needs regardless.