Added Input System Package, Interaction System, Door Switch, Wall Running
Added Input System Package as old is outdated and newer is recommended, will need to add more inputs down the line.
Interaction System is now implemented albeit at an early stage. It works off raycast from camera as origin both as position and origin to a few units. Need to add an input for interact as currently it fires on each cast hit.
Door Switch was implemented as a way of testing interaction. Made an interaction interface which will be cast to and fired when the interaction button is pressed in future versions (not implemented that part yet).
Wall Running now applies gravity again albeit at half-rate.
Also updated GitIgnore to exclude .idea folder made as a result of using Jetbrains rider.