Shooting + Wall Run work
Wall running is closer to where it should be. Now the movement velocity while running is grabbed from the player on start of run to prevent bug where player goes other side of wall after run.
Basic version of projectile shooting is implemented with pooling. Need to have a pool cleanup function when deleted as enemies should use pooling also.