Pooling Improved
Pooling now doesn't need a manager object to be pre-created and now checks if one exists before creating if not. As there is only one projectile based gun in this game this is fine without checking if the pool matches the prefab being spawned but this check is easy enough to add in if so.