Check Point System Works
Checkpoints now work properly. I had to do a sort on the list of checkpoint objects and now they appear and disappear when picked up. Currently they have a cool explosion/smoke effect on them but this will be changed later.
Problem initially was the trigger being attached to a child wasn't firing the function.
The other issue with the player being reset to random points was the fault of the list being unsorted which has now been corrected using OrderBy function comparing the IDs.