Basic Sprint System Implemented.
Using the CallbackContext on InputAction to ensure it fires once. The sprint button is now a toggle, when sprinting movement speed is doubled (setting via hardcoded values as apposed to flexibly doubling at this time).
At the moment inside the GetMovementInput function it also checks the magnitude of the move vector on the InputAction mapping and if it's less than .1 it starts the stop sprint coroutine where if uninterrupted for .3 seconds it stops the sprint otherwise if the magnitude goes above .1 in that time then the coroutine is stopped.