Switched to Behavior Designer plug-in
Replaced TheKiwiCoder AI Tree with Behavior Designer from the Unity Store. This plug-in is a lot easier to work with than the prior free asset and enables debugging a lot easier so it's much more suited for my use.
Updated AIPerception to work better with this plug-in, the new trees can read from other scripts so I just use a public bool for SeePlayer. Also fixed the bug where the perception is too high to detect the player and now it's origin is set by an empty transform on the NPC.