Now Presets Variables on the AIPerception Script (soon to be renamed) + Now sets blackboard directly as apposed to referencing Perception bools
AIPerception will soon be renamed to AIController as it now fits a more general purpose by also setting up the BB as well as Perception. (Fixing this now as this will be required for ease-of-use especially with multiples of each enemy type (I also learnt how to do this recently).
Setting Blackboard via script is also a much better solution due to cleanliness etc so I have made a function that does this also on the AIPerceptionScript called ReportBoolToBlackboard which allows me to type the name of string and also the value and then reports that, no error checking in place but if in a larger solution that'd be necessary.